Fetch blog post in layout not header
